requirements.txt
#-e .
numpy
pandas
scikit-learn
python-dotenv
mysql-connector-python
pymysql
SQLAlchemy
seaborn
matplotlib
catboost
xgboost
mlflow



